Move the error unwrapper into common (#1558)
