+++ packages/common/src/types.ts
@@ -11,10 +11,8 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-export * from "./utils.js";
-export * from "./error.js";
-export * from "./logger.js";
-export * from "./array.js";
-export * from "./hash.js";
-export * from "./string.js";
-export * from "./types.js";
+export type ApiJsonError = {
+	message: string;
+	key?: string;
+	code: number;
+};
