+++ packages/locale/src/locales/en.json
@@ -108,7 +108,10 @@
 		"INCORRECT_CAPTCHA_TYPE": "Incorrect CAPTCHA type",
 		"INVALID_SITE_KEY": "Invalid site key",
 		"UNAUTHORIZED_ORIGIN_URL": "Unauthorized origin URL",
-		"INVALID_IP": "Invalid IP"
+		"INVALID_IP": "Invalid IP",
+		"INVALID_URL": "Invalid URL",
+		"INVALID_BODY": "Invalid body",
+		"PARSE_ERROR": "Error parsing request"
 	},
 	"CLI": {
 		"PARAMETER_ERROR": "Invalid parameter"
