+++ packages/locale/src/locales/es.json
@@ -108,7 +108,10 @@
 		"INCORRECT_CAPTCHA_TYPE": "Tipo di CAPTCHA errato",
 		"UNAUTHORIZED_ORIGIN_URL": "URL de origen no autorizada",
 		"INVALID_SITE_KEY": "Clave de sitio no válida",
-		"INVALID_IP": "IP no válida"
+		"INVALID_IP": "IP no válida",
+		"INVALID_URL": "URL inválida",
+		"INVALID_BODY": "Cuerpo inválido",
+		"PARSE_ERROR": "Error al analizar la respuesta"
 	},
 	"CLI": {
 		"PARAMETER_ERROR": "Parámetro inválido"
