+++ packages/locale/src/locales/fr.json
@@ -108,7 +108,10 @@
 		"INCORRECT_CAPTCHA_TYPE": "Type de CAPTCHA incorrect",
 		"UNAUTHORIZED_ORIGIN_URL": "URL d'origine non autorisée",
 		"INVALID_SITE_KEY": "Clé de site non valide",
-		"INVALID_IP": "IP invalide"
+		"INVALID_IP": "IP invalide",
+		"INVALID_URL": "URL inválida",
+		"INVALID_BODY": "Cuerpo inválido",
+		"PARSE_ERROR": "Erreur d'analyse de la demande"
 	},
 	"CLI": {
 		"PARAMETER_ERROR": "Paramètre invalide"
