+++ packages/locale/src/locales/it.json
@@ -108,7 +108,10 @@
 		"INCORRECT_CAPTCHA_TYPE": "Tipo di CAPTCHA errato",
 		"UNAUTHORIZED_ORIGIN_URL": "URL di origine non autorizzato",
 		"INVALID_SITE_KEY": "Chiave del sito non valida",
-		"INVALID_IP": "IP non valido"
+		"INVALID_IP": "IP non valido",
+		"INVALID_URL": "URL non valido",
+		"INVALID_BODY": "Corpo non valido",
+		"PARSE_ERROR": "Errore nell'analisi della richiesta"
 	},
 	"CLI": {
 		"PARAMETER_ERROR": "Parametro non valido"
