+++ packages/locale/src/locales/pt-BR.json
@@ -108,7 +108,10 @@
 		"INCORRECT_CAPTCHA_TYPE": "Tipo de CAPTCHA incorreto",
 		"UNAUTHORIZED_ORIGIN_URL": "URL de origem não autorizada",
 		"INVALID_SITE_KEY": "Chave de site inválida",
-		"INVALID_IP": "IP inválido"
+		"INVALID_IP": "IP inválido",
+		"INVALID_URL": "URL inválido",
+		"INVALID_BODY": "Corpo inválido",
+		"PARSE_ERROR": "Erro ao analisar a resposta"
 	},
 	"CLI": {
 		"PARAMETER_ERROR": "Parâmetro inválido"
