+++ packages/provider/src/api/errorHandler.ts
@@ -12,11 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 // We need the unused params to make express recognise this function as an error handler
-import { type ProsopoApiError, ProsopoBaseError } from "@prosopo/common";
-import { i18n as i18next } from "@prosopo/locale";
-import type { ApiJsonError } from "@prosopo/types";
+import { type ProsopoApiError, unwrapError } from "@prosopo/common";
 import type { NextFunction, Request, Response } from "express";
-import { ZodError } from "zod";
+import type { ZodError } from "zod";
 
 export const handleErrors = (
 	err: ProsopoApiError | SyntaxError | ZodError,
@@ -30,37 +28,4 @@ export const handleErrors = (
 	response.status(code);
 	response.send({ error: jsonError });
 	response.end();
-};
-
-export const unwrapError = (err: ProsopoApiError | SyntaxError | ZodError) => {
-	const code = "code" in err ? err.code : 400;
-	let message = err.message;
-	let jsonError: ApiJsonError = { code, message };
-	let statusMessage = err.message;
-	jsonError.message = message;
-	// unwrap the errors to get the actual error message
-	while (err instanceof ProsopoBaseError && err.context) {
-		// base error will not have a translation key
-		jsonError.code =
-			err.context.translationKey || err.translationKey || jsonError.code;
-		jsonError.message = err.message;
-		if (err.context.error) {
-			err = err.context.error;
-		} else {
-			break;
-		}
-	}
-
-	if (err instanceof ZodError) {
-		message = i18next.t("CAPTCHA.PARSE_ERROR");
-		statusMessage = message;
-		if (typeof err.message === "object") {
-			jsonError = err.message;
-		} else {
-			jsonError.message = JSON.parse(err.message);
-		}
-	}
-
-	jsonError.code = jsonError.code || code;
-	return { code, statusMessage, jsonError };
 };
